External stylesheet for sold listings map and sidebar
Prior to this release, the custom styling for the Sold Listimgs Map page was stored in the Squarespace page header section on the page. This release moves the CSS to an external stylesheet file for scalability across multiple pages. And the CSS was added to a separate CSS file (instead of site.css) due to potential class conflicts.