First release of token replacement javascript
Javascript that dynamically replaces tokens with page-specific values, e.g., [PAGE_TITLE]. For this release, the script is called from the page settings in whith the token replacement is implemented on the page; instead of the site global implemention. This limited scoping is done to minimize the CSS-driven page flicker on loading.